modules/legal-ingest/requirements.txt
# Core dependencies
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.1
python-dotenv==1.0.1

# HTTP and scraping
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.2.2

# PDF processing
pypdf==4.2.0
# pytesseract==0.3.10  # Uncomment for OCR support

# Data processing
PyYAML==6.0.1

# Database connectors
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Search and embeddings
# opensearch-py==2.5.0  # Uncomment when OpenSearch is configured
# pymilvus==2.4.0  # Uncomment when Milvus is configured
sentence-transformers==2.7.0

# Monitoring
prometheus-client==0.20.0

# Workflow orchestration (optional)
# apache-airflow==2.9.0  # Install separately if using Airflow

# Utilities
numpy==1.26.4
